Fix chipmunk audio: sound_compute_random_scale call site read float tag fields as int
The prototype was correctly updated to all-float params, but the caller
at sound_manager.c:1933 still cast tag offsets +0x14/+0x18 as *(int*).
The compiler generated int-to-float conversion on the IEEE bit patterns,
producing ~1e9 scale values that spiked every sound's pitch.

+++ src/halo/sound/sound_manager.c
@@ -1930,8 +1930,8 @@ int sound_start(int sound_tag_index, void *source, int object_handle,
         /* Compute random scale for this sound instance. */
         {
           float rscale =
-            sound_compute_random_scale(*(int *)((char *)sound_tag + 0x14),
-                                       *(int *)((char *)sound_tag + 0x18),
+            sound_compute_random_scale(*(float *)((char *)sound_tag + 0x14),
+                                       *(float *)((char *)sound_tag + 0x18),
                                        *(float *)((char *)sound_tag + 0x44),
                                        *(float *)((char *)sound_tag + 0x5c),
                                        *(float *)((char *)source + 4));
